requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.2
Pillow==9.2.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
python-dotenv==0.21.0
sqlparse==0.4.3
toml==0.10.2
